Remove oidc and add login, use sqlite

+++ kleiderkammer/kleidung/model/kleidungsleihe.py
@@ -8,7 +8,7 @@ Base = declarative_base(cls=DeferredReflection)
 
 class Kleidungsleihe(db.Model):
     __tablename__ = 'kleidungsleihe'
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     mitglied_id = db.Column(db.BigInteger, db.ForeignKey('mitglied.id'))
     kleidung_id = db.Column(db.BigInteger, db.ForeignKey('kleidung.id'))
     von = db.Column(db.DateTime, nullable=False)
